Admin logs for 3rd Party Victory events (#5658)
# About the pull request

Adds logging in the event that W-Y, UPP, or CLF win during a xeno minor.

# Explain why it's good for the game

Would be good to track how common (or rare) these sorts of things are,
which would also be good feedback on if its worthwhile to make more
faction victory moments.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
admin: added logging for 3rd party victory events
/:cl: