Fixes trash carts not working with fultons (#5664)
# About the pull request

Fixes #5621, which was caused by a runtime in
`/obj/structure/dropship_equipment/fulton_system/proc/automate_interact()`.
The cart's entry in `possible_fultons` has a key of `"�\u{16}trash
cart"`, but the `fulton_choice` passed from the UI is just `"trash
cart"`. Because they don't match, `fult` is unable to be set, and an
error is thrown on L1154.

https://github.com/cmss13-devs/cmss13/blob/2657502ad1c5de88fd0aef35072900a736be84fe/code/modules/cm_marines/dropship_equipment.dm#L1141-L1154

# Explain why it's good for the game

Bugfix!

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

**Before:**
Runtime error. *(I can't be bothered to go through the set up again just
to record this)*

**After:**


https://github.com/cmss13-devs/cmss13/assets/57483089/9ee496a4-d201-43b2-95d2-963dcc7c8569

</details>


# Changelog
:cl:
fix: Fixed trash carts not being pick-uppable by the dropship's fulton
recovery system.
/:cl: