Adds One ACCESSORY ONLY Slot for Helmets (#5573)
# About the pull request

Allows for the insertion of an extra cosmetic item in helmets that
subtype the M10.

# Explain why it's good for the game

Adds more personalization to people who don the helmet. Change is purely
cosmetic.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Stakeyng
qol: Lets you put an extra accessory item in your helmet.
/:cl: