Thour tiny tutorial tweaks (and a unit test) (#5695)
# About the pull request

<ol>
    <li>
Commit 5014927bc4ae6f3bd03f3ea1866cb6ba6c9e7c54: Makes the 'Tutorial
Menu' automatically close when a tutorial is started.
        <ul>
<li>Also adds an <code>abort_tutorial()</code> call to
<code>/datum/tutorial/proc/start_tutorial()</code>, since it may have
been possible to for player to get trapped in the tutorial
otherwise.</li>
        </ul>
    </li>
<li>Commit b10974c0471ef14c3c66ec10910e8c993cbb29d1: Adds a
<code>tutorial_id</code> to the basic Xenomorph tutorial.</li>
<li>Commit 38c03c987dbad50f7d8a2354a63c75f88d5efcce: Edited the tutorial
Xenomorph a bit so that it's always "You are Inquisitive Drone
(XX-123)", as opposed to "You are Inquisitive Young/Mature/etc. Drone
(XX-123)".</li>
    <li>
Commits a595b971c489922510bd984e3d324ac844450efe +
27e38e654a611aeefd548c41035bbff035caa76b: Adds a basic unit test for
<code>/datum/tutorial</code> subtypes to verify that they have set some
basic variables.
        <ul>
<li>Also lightly organises the <code>#include</code>s in
<code>code/modules/unit_tests/_unit_tests.dm</code>.</li>
        </ul>
    </li>
</ol>

# Explain why it's good for the game

Just a couple of things I noticed while exploring tutorial code.
The third commit is maybe a bit subjective, but I personally think it
makes the tutorials seem like more of a separate game state, if that
makes sense.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
ui: Made the Tutorial Menu automatically close when a tutorial is
started.
code: Made the tutorial Xenomorph not inherit the user's 'age' prefix.
code: Added a basic unit test for tutorials.
/:cl: