Revive Lesser Drone and Facehugger temporary communication restrictions (#5688)
# About the pull request

This PR revives and expands #5352. By expand I mean that in addition to
a 3 minute restriction for say and hivemind speak, custom emotes and
pointing are also restricted during the 3 minute window. Past that I did
some additional refactoring.

# Explain why it's good for the game

As a ghost role, this acts as a measure to mechanically limit the
communication of meta info to other members in the hive. For now the
restriction is only 3 minutes.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/fd64c2f4-9142-4fe1-9953-b29c480c17ee)

</details>


# Changelog
:cl: Zonespace Drathek
balance: Lesser drones and Facehuggers cannot speak, custom emote, or
point for 3 minutes after spawning
/:cl:

---------

Co-authored-by: John Doe <gamingskeleton3@gmail.com>