Fixes crates layering under the fulton recovery system (#5673)
# About the pull request

Fixes crates and some other fultonable objects being on a layer below
the dropship's fulton recovery system.

This was happening because crates (any closet subtypes really) are on
`BELOW_OBJ_LAYER`, and dropship equipment is on `ABOVE_OBJ_LAYER`.
Changing the layer of `/obj/structure/dropship_equipment` to something
lower would probably work, but I felt the safest solution would be to
just set the equipment's `plane` to whatever the attach point has.
This has no effect on the outside attachment points, but fixes the
layering issue since the inside ones are on `FLOOR_PLANE`.

I should also mention that the actual fix is in the first commit
(b8e44956cfb4ed11d971456be8059fd9eaeb928f) and is only one line. All the
rest of this is just because the `install_equipment()` proc was a bit
confusing as it was.

# Explain why it's good for the game

Unless I've missed something in git blame, this has probably been a bug
for 4+ years.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

**Before:**

![image](https://github.com/cmss13-devs/cmss13/assets/57483089/77d8e10c-6354-4902-a1c1-982a69709612)

**After:** (Also a regular crate too)

![dreamseeker_htRSczUHZh](https://github.com/cmss13-devs/cmss13/assets/57483089/a7a93420-c7c0-4203-b4df-c3f3415cc98c)
![dreamseeker_yV0pi23XYY](https://github.com/cmss13-devs/cmss13/assets/57483089/8e4155ee-5b5e-437a-9cde-788755cc164a)

**Verifying that other hardpoints aren't affected:**

![ds](https://github.com/cmss13-devs/cmss13/assets/57483089/1e67da09-910a-452e-ba26-27df763d0a31)

</details>


# Changelog
:cl:
fix: Fixed crates and similar objects layering below the dropship's
fulton recovery system.
/:cl: