Bipod now gives you an option to switch to full auto when deploying (#5638)
# About the pull request

Attaching the bipod to weapons with full auto capability or firearms
marked with the GUN_SUPPORT_PLATFORM flag will give the user a button
toggle to immediately switch to full auto when deploying the bipod.


![image](https://github.com/cmss13-devs/cmss13/assets/17518895/cea29ccc-5761-4a7b-99bd-cbdd035bc52b)



If the full auto switch is turned on, the weapon will simply switch to
full auto when the bipod is deployed. If you undeploy it, it will switch
back to whatever previous firemode the weapon was in, e.g:

1. You set your pulse rifle to semi auto. 
2. You deploy the bipod. 
3. The pulse rifle is set to full auto. 
4. You undeploy the bipod. 
5. The pulse rifle will switch back to semi auto.

In cases where you switch the weapon's firemode when it's already
switched to full auto, it won't revert back to whatever firemode was set
before it was deployed, e.g:

1. You set your pulse rifle to semi auto. 
2. You deploy the bipod. 
3. The pulse rifle is set to full auto. 
4. You set your pulse rifle to burst fire.
5. You undeploy the bipod. 
6. The pulse rifle will stay on burst fire.

Also renames a couple of one letter variables.

# Explain why it's good for the game

Quality of life. Less messing around with firemode selectors which is a
terrible headache, especially for the HPR, which you usually want to
have on full auto 90% of the time. Those who want to stick with the old
way don't have to change anything, as the full auto switch is entirely
optional.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/cmss13-devs/cmss13/assets/17518895/35bcc6e7-947d-4f82-971f-ac22b723537b



</details>


# Changelog
:cl:
qol: Attaching the bipod to weapons that have full auto (HPR included)
will give you a button toggle for immediately switching to full auto
when deploying the bipod.
/:cl: