Added two new beacons, a mug and a suicide pill to the CLs briefcase  (#5627)
# About the pull request

Added a cyanide pill to the CLs briefcase when the MPs are breaking down
your door and you don't want to rot in jail for growing a xeno hive in
their backyard.

Added a Weyland-Yutani coffee mug to the CLs briefcase for more BRAND
ADVERTISING!

Added two new beacons for the CLs briefcase.

### **First**
Bodyguard Beacon. Send a message to PMC Dispatch and request a single
Corporate Security Officer for any of your emergency needs.

### **Second**
Lawyer Beacon. Got a CO that broke the contract he signed seconds
before? Got a survivor who spoke to the CC and broke his NDA? Corporate
Affairs presents you with a beacon, aiming to solve all your contract
and Corporate needs with a specialised duo of expert Lawyers!


# Explain why it's good for the game

Taking inspiration from this forum post on ideas guys, I found some very
interesting ideas and decided to put the peoples thoughts in a PR.


https://forum.cm-ss13.com/t/cl-briefcase-and-what-items-you-want-to-see/6268

The cyanide pill can be used if in a pickle and the USCM is on your ass
for some ML violation. Or some sort of W-Y suicide pact if you have more
loyalists to your cause.

The coffee mug is a nice souvenir, why not you know. It's there in the
code, make it have some use.

### **The Big Ones**

The two beacons, maybe some people have issues with this one. The
bodyguard and lawyer one BOTH cost 50 points each (the max) as to not be
abused and bought every minute.

Let me begin with saying, these work like the CMB/PMC Beacons the
Marshal/Goon Professor have, meaning they need **ADMIN PERMISSION TO BE
SENT.** If an admin decides to allow either one of these beacons then
the first one has a single Corporate Security Officer, incase the CL is
threatened and the MPs are inactive, or when there's a person breaking
in the CLs office, or any other scenarios which the Executive would need
Company protection. A SINGLE Corporate Goon I feel will not impact the
round as much or cause any extra havoc.

Secondly, the two man Lawyer team which spawn with notepads, pens,
clipboards, tape recorders, the lead lawyer with a secure briefcase and
1000 cash counterfit. They have no guns, no bodyguards, they are
basically a slightly better version of the CL, or slightly worse. And it
can increase the RP? Maybe make the CO return from deployment? Harrass
the Captain long enough so he orders your arrest? Who knows, we'll find
out.


# Changelog
:cl:
add: Added two new beacons for the CL, Lawyer and Bodyguard.
add: Added an Executive Specialist and Executive Supervisor Lawyer
preset.
add: Added a cyanide pill in the CLs briefcase for an easy way out.
add: Added a Wey-Yu Coffee Mug in the CLs briefcase as a souvenir sale.
/:cl:

---------

Co-authored-by: Jeff Watchson <johncenatoutourourou@gmail.com>