Xeno strains refactor (#5542)
# About the pull request

Refactors xeno strains by taking out the old
`/datum/xeno_mutator`/`/datum/mutator_set` system and replacing it with
a new `datum/xeno_strain`, since as far as I'm aware mutators haven't
been touched in almost 5 years.

As a side note, the DMI changes are from me deleting the icon states of
the unused 'Tremor' Burrower strain, since I figured it fit within the
bounds of the PR. (The strain's sprites are identical to that of the
base caste.)

# Explain why it's good for the game

The new datum should be a lot more streamlined than the old system, and
require much less boilerplate when making a new strain. Ideally this
could even replace `/datum/behavior_delagate` with signal handlers in
the future, although a few more refactors would be required before
that's possible.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
refactor: Refactored xeno strains.
add: Added a keybind setting to open the 'Purchase Strain' window. (Also
renamed the macro)
/:cl: