Tacmap ability fix
Not actually caused by the PR, just discovered by it. (hooray I guess?)

Basically when the tacmap action was removed by a strain (with a new one added later), the `COMSIG_HIVE_NEW_QUEEN` signal on it never got unregistered and it sorta got left floating in nullspace until a new queen comes along to cause a runtime.
There should really be a unit test for this problem, but that's out of scope for now.

(Although I did pre-emptively fix /datum/action/minimap in case it ever has the same problem)