Increases RMCs/TWE's Rifles ROF (#5684)
# About the pull request

Increases the ROF of the RMC's rifles to be comparable to the M41As. 

# Explain why it's good for the game

The RMC's guns are simply non-viable by ERT standards let alone Marine
standards. Their ROF is too low to be an effective damage dealer in
combat. This, combined with the low amount of ammo RMC's get, often
leads to RMC's ditching their unique guns for a M41A. The only advantage
of the guns is they don't inhibit movement speed, but I do not believe
this is sufficient of a boon to justify the terrible ROF.

Furthermore, ERT guns are generally slightly better, or far superior, to
USCM weapons due to their limited nature and the small numbers of spawns
that are present. Ergo, this buff is not egregious compared to UPP, PMC
or other ERT spawns. Furthermore, the RMC are described as a tech
advance, if small, commando force, it is justifiable that elite
commandos get access to slightly better guns than the rank and file.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Royal Marine Commando's firearms now have a higher rate of
fire, comparable to the M41A.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>