Fix bug with RestrictedInput TGUI component (#5651)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Should fix bug in restricted input TGUI component. The number was being
clamped on every input, which meant that applying a number of 30 into a
range of 19-90, 3 would be bounded to 19 and then the 0 would make the
value 190, being bounded to 90.

This change performs the check on blur.

# Explain why it's good for the game
Fixes https://github.com/cmss13-devs/cmss13/issues/5610

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
code: rework logic of restrictedinput component to reduce checks
/:cl: