Vamp lurker Tweaks (#5625)
# About the pull request


-Using Tail stab on the same tile a enemey is standing will
automatically target tailstab to the enemy.
-Sprite clicking and enemy with Tail stab now provides a Bonus 15 Dmg

# Explain why it's good for the game

after the reversal of the vamp lurker rework it was left in a very weak
state . only a handfull of players being able to play it decently. this
PR aims to make it a little easier for players to use vamp lurker.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Tail Jab will now target enemies on the same tile it was
targeted.
balance: Aiming Tail Jab directly on a target will do an additional 15
damage
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>