adding the ability for SG to buy the VP78 and the SU6 (#5701)
# About the pull request
adding the ability for SG to buy the VP78 and the SU6
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
it's clear that SG is suppose to have the option to have a sidearm so i
am giving him the ability to by the VP78 and the SU6
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: add the ability for SG to buy an VP78 or an SU-6 each for 15
points.
/:cl:

Co-authored-by: Julien <jverger.ingx@gmail.com>