Temporary mute now also prevents hive death messages (#5748)
# About the pull request

This PR is a followup to #5688 to additionally prevent death messages
getting sent to the hive during the temporary mute time.

# Explain why it's good for the game

Ultimately this is to quell more concerns with ghosts using meta
knowledge of say a flank approaching and throwing themselves at that
flank and informing the hive of the flank via their death message
location.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/cmss13-devs/cmss13/assets/76988376/8b8426bd-8f82-43f6-bb70-a91346c47534

</details>


# Changelog
:cl: Drathek
del: Removed death messages for lesser drones and facehuggers that are
still temporarily muted
/:cl: