Adds tooltips to changelog icons (#5713)
# About the pull request

Adds tooltips to the changelog icons.

# Explain why it's good for the game

While some of the icons are easy to understand at a glance, others like
those for 'qol', 'refactor', 'mapadd' and 'config' are pretty much just
meaningless glyphs. They work fine if you've also read the discord
changelog or checked recent github merges, but if you're just going off
of the ingame one it can sometimes be a little confusing.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/cmss13-devs/cmss13/assets/57483089/cc7935e6-da77-4c1f-b2e9-ab38c37aaaf9

</details>


# Changelog
:cl:
ui: Added tooltips to entry icons in the changelog.
/:cl: