Adds holo-targetting Vulture rounds (#5719)
# About the pull request

Adds holo-targetting rounds for the Vulture. Their characteristics are:

- They deal SIGNIFICANTLY LESS damage, down from 400 to 60.
- They cannot penetrate walls nor mobs.
- They do not slow down mobs that are hit.
- When they hit a target, they will "mark them for death", inflicting a
holo stack value of 333. This means that they will take 33% more damage
from all sources. A balloon alert will be shown to all nearby mobs
indicating the mark, along with the holo-target aura itself.
- **They have IFF.**


![image](https://github.com/cmss13-devs/cmss13/assets/17518895/69fc22e2-316d-43b5-87fd-08e2a5e9ba2b)


The bullet increases the target's max capacity for holo stacks increases
by 233 (to 333 in total). Any holo-targetting rounds that hit the mob
will be able utilize this cap increase. The holo stacks will completely
deplete after 35-ish seconds from the initial impact. It takes 5 seconds
for the stack drain to begin, with every second passing after that
removing 10 stacks. Other holo-targetting rounds will not modify the
rate of stack depletion, keeping it at a firm 10 stacks per second. They
will, however, stop the drain for 5 seconds.


![image](https://github.com/cmss13-devs/cmss13/assets/17518895/ad85d159-344b-407a-8158-23eafbf0824a)


# Explain why it's good for the game

This gun is really fun for the users, encouraging a buddy system that I
feel the game desperately needs more of. I will admit that being hit
with 400 damage, 50 AP from 2 screens away is not engaging for the
opposing side, giving them no counterplay at all.

With these bullets, the Vulture can take a supporting role for the
marines instead of being an off-screen death cannon. 33% is a decent
damage increase, but it requires multiple marines around the target to
truly take advantage of it. Compared to the payoff that a regular
Vulture round gives, the lack of IFF for those bullets is
understandable. However, given that you NEED friendly troops to get any
use out of these rounds, I've given the rounds IFF (with Zonespace's
blessing) to broaden the areas and situations where it could get used.

The ammo is still limited as ever, so the sniper team will have to pick
their shots. Given how weak the rounds are in comparison to the standard
ones, they should rarely be able to snipe off targets on their own (in a
rare circumstance that they spot a half-crit xeno, an M4RA with a 4x
scope can finish them off just the same).

Is it weak? Too situational? Maybe, but the weapon is very interesting
mechanically and I would just like to see it in play a bit more.

The weapon is still ADMIN-ONLY. This will only give the admins a less
blunt variant of the Vulture if they choose to spawn in it.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

How being hit feels:


https://github.com/cmss13-devs/cmss13/assets/17518895/a576629e-7961-4421-9932-d707e12477f1

General test:



https://github.com/cmss13-devs/cmss13/assets/17518895/e2fbb117-1cf9-4945-b87a-d63e68f62a10

Bonus damage testing. The M41A does a base damage of 44. A holo-stack of
333 will result in roughly 44 * 1.33 = 59 damage.


![image](https://github.com/cmss13-devs/cmss13/assets/17518895/d230fc11-34e6-468e-89c3-394696c9b3e1)


</details>


# Changelog
:cl:
add: The Vulture can now be loaded with holo-targetting rounds which
have severely hampered ballistic performance, but mark any target that
is hit, increasing their damage taken by 33% for a brief period of time.
/:cl: