Removes VAIMS backpack burn line, adds extra mag. (#5670)
# About the pull request
VAIMS (mil contractor medical specialist) had a redundant burn line
(because surgical webbing spawns with burn line). this PR removes it and
replaces it with an MAR mag to fill the empty space.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
the burn graft is a waste of space in the backpack. As for the MAR mag I
put it in so that the inventory doesn't have a random open space, i'm
fine removing it if a maintainer tells me to.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
del: VAIMS (vaipo med spec) no longer gets a redundant burn line
balance: VAIMS now gets an extra MAR mag.
/:cl: