Allows the MOU53 to be stored in the shotgun scabbard. (#5705)
# About the pull request

Lets you store the MOU in the shotgun scabbard.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Shotgun scabbard is soul. Basically no advantage given, other than the
fact that you can keep a slot reserved for your MOU.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Stakeyng
add: mou is now compatible with the shotgun scabbard
/:cl: