Food processor can be used to remove carpotoxin and acidic blood with master domestic skill (#5577)
# About the pull request

This PR makes it so that characters with master level domestic skill are
able to remove carpotoxin and acidic blood from carp fillets and
xenomeat. Xenomeat, carp fillets and recipes including those have had
their levels adjusted to avoid adding additional carpotoxin/xenoblood
when creating a recipe so a processed ingredient won't result in a
dangerous recipe.

I've tested this using characters with and without the skill with flour,
potatoes, xenomeat and carp fillets to ensure that only the affected
processor uses are changed, that processing works correctly and that the
skill is required.

# Explain why it's good for the game

This would open more opportunities for the mess technician to serve the
crew with recipes that rarely see the light of day due to being
dangerous (while leaving some risk if an MST forgets to process the
ingredients). Some marine players might also enjoy the roleplay
opportunities around getting the chef to turn their fallen enemies into
burgers.

# Testing Photographs and Procedure
<details>

<summary>Screenshots & Videos</summary>

Here is shown a character with insufficient skill unable to modify the
food, before putting it in the soda fountain to show its reagents:

https://github.com/cmss13-devs/cmss13/assets/32908370/8138d1c5-77a2-4f7d-92e9-9d7059fa3f05

Here is shown a character with sufficient the skill able to modify the
food, before putting it in the soda fountain to show its reagents:

https://github.com/cmss13-devs/cmss13/assets/32908370/93bb0c99-fccf-46ca-a786-579be203629b

</details>


# Changelog
:cl:
add: Increased levels of acidic blood and carpotoxin in xenomeat and
carp fillets.
del: Xenomeat and carp fillet-based recipes no longer add additional
acidic blood or carpotoxin.
qol: Master level domestic skill allows the use of the food processor to
remove acidic blood and carpotoxin from xenomeat and carp fillets.
fix: Turning named pieces of human meat into meatballs will retain the
human's name.
/:cl:

---------

Co-authored-by: SabreML <57483089+SabreML@users.noreply.github.com>