refactor barricades files to hopfully have a more constant interaction when trying to fix them with welder. (#5609)
# About the pull request
fixes: https://github.com/cmss13-devs/cmss13/issues/5229
1-replace a bunch of single letter variable and in the same movement
standardize it for all the parent/child object...
2-create a proc that is a group of check to see if you can start fixing
a barricade with a welder.
3-implement this proc on metal,plasteel and deployable to have all the
check be the same as we want them to all react the same way...
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
i tested welding different different cades at the same time seem to work
i had some weirdness but it's working.
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
refactor: refactored files with the aim of making fixing barricade with
a welder more constant.
fix: being unable to repair more than one metal barricade at the time.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>
Co-authored-by: Birdtalon <birdtalon@gmail.com>