Abomination rework and predator plasma rifle buff. (#5432)
# About the pull request
This pr intends to catch up predalien towards current gameplay by
reworking a few abilities and adding new ones.
# Explain why it's good for the game
Currently the predalien has been untouched (practically) for over 3-4
years now, while being rare and decent enough to hold its own against
current year marines and predators, its highly outdated. It's main power
being "the gib" which is not fun for marines to be instantly removed
from the game as most of the time you will be accompined by a queen and
the entire hive. its not being fun being comboed into a queen screech +
gib

As their natural counter, the predators are currently lacking in a
strong way to counter it, i believe the Plasma Rifle thats only
available when an abomination is present is very weak, underpowered and
generaly not that good. I can go into deeper details if anyone wants me
to. So my natural response was to make it set people on fire because
generally the predalien will be guarded by an entire hive. This will
make it so that atleast the xenomorph hit, will have to back off
allowing the predator to breathe. I havent touched the recharge rate
because i think this incin bullet will be powerful enough to compensate
for the low recharge rate, and only 16 shots (you need 14 dead on shots
to kill an abomination, not counting its movement, phermones, and the
entire hive guarding it.)

below are all the changes ive decided to make;
"Feral Frenzy" is a mashup of the old gut ability, it now has a toggle
that lets you choose between AOE/SINGLE TARGET. I think this is a
defining factor in predalien reworks, because currently it is too
focused on fighting predators and not enough on marines. This will help
you fend off against a marine group by letting you become devastating
against grouped up marines. Or chosing to catch one by switching to
single target since it has a root. (Also retaining its ability to fight
off predators).

"Feral Rush"; Feral rush incrases your armor and speed by a certain
number, i wanted to add this instead of the useless pounce as i felt
like the pounce was basically not providing you with anything other then
just a distance closer, and it didnt fit in with the new gutting
abilities, this way you can speed up, armor up and then go inside the
marine force and do your aoe ability and get back out. This will ensure
that you arent useless while going in since without the armor and speed,
you will melt by the time you go in marine numbers, (also remember the
AOE gut has a windup of 2 seconds.), The armor is removed 6 seconds
after the slow (the speed lasts 3 seconds, so you're just armored for an
extra 3 seconds) as to have that extra armor while retreating from more
open places.
"Stomp" ; the reason ive decided to remove this ability is it felt way
out of place for this kit and just simply wasnt good enough (in my
opinion) to being there although a slow or aoe root might be needed if
this gets approval enough for gametesting.

"Frenzy Smash" ; Grabs somebody by the leg and smashes them to the
ground, this honestly just felt like a cool ability to add, it lets you
keep somebody in place but also keep them moving, this is mostly a flair
ability but still VERY deadly, may be buffed to make preds slower for a
duration. deals 20 damage and scales with kills
 

THIS IS ALL UP FOR DEBATE, AND I WOULD LIKE TO WATCH THIS ALL INGAME VIA
ABOM EVENTS BY ADMINS IF THATS POSSIBLE.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/cmss13-devs/cmss13/assets/47158596/f28a5048-701f-4b9d-a2bf-dd5dbc6d08be


https://github.com/cmss13-devs/cmss13/assets/47158596/53da1e18-1a4a-49de-a73d-1ab7f1abb1d8


https://github.com/cmss13-devs/cmss13/assets/47158596/69eeb0e9-cbe6-4712-a15a-b599dbc56109




</details>


# Changelog


:cl:
add: Added a new ability to the predalien "Feral Rush" that increases
it's armor and speed for a short duration.
add: Added two new abilities to the predalien "Feral Frenzy" and a
toggle. The predalien can now switch between a single target GUT and an
AOE one, both of which has damage that scale with your kills.
del: Removed the predaliens gib ability
del: Removed the predaliens "pounce" ability
del: Removed the predaliens "Ground Smash" ability
balance: Predalien no longer has plasma costs, or plasma. 
balance: The Plasma Rifle, which is ONLY used to hunt abominations now
has a higher ROF and has incin bullets.
spellcheck: Re-wrote the predalien text to be more up to date and remove
missinformation
qol: Everyone can see how many kills the predalien has by examining it.
add: Added a "Feral Smash" ability that lets you grab somebody and smash
them to the ground, this scales with kills and is a devastating attack.
balance: Removed screenshake from predalien's screech
balance: Predalien removes fire stacks faster.
fix: Girders are now slashable by very_sharp_claws instead of just
having a queen chack
/:cl:

---------

Co-authored-by: InsaneRed <prodexter31@outlook.comm>
Co-authored-by: Birdtalon <birdtalon@gmail.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>